+++ src/main/java/huce/edu/workmanagementprojectbackend/controller/EmployeeController.java
@@ -1,6 +1,5 @@
 package huce.edu.workmanagementprojectbackend.controller;
 
-import huce.edu.workmanagementprojectbackend.model.DepartmentEntity;
 import huce.edu.workmanagementprojectbackend.model.EmployeeEntity;
 import huce.edu.workmanagementprojectbackend.services.department.IDepartmentService;
 import huce.edu.workmanagementprojectbackend.services.employee.IEmployeeService;
@@ -33,6 +32,7 @@ public class EmployeeController {
                                         HttpSession session){
     model.addAttribute("employees",iEmployeeService.getAll());
     model.addAttribute("user",session.getAttribute("user"));
+    sentError(model);
     return "/html/Manager/employee/manager-employee";
   }
 
@@ -41,20 +41,29 @@ public class EmployeeController {
                                     HttpSession session){
     model.addAttribute("departments",iDepartmentService.getAll());
     model.addAttribute("user",session.getAttribute("user"));
+    sentError(model);
     return "/html/Manager/employee/manager-employee-add";
   }
 
   @PostMapping("/save_employee_action")
   public String addEmployee(@ModelAttribute("employee") EmployeeEntity employee){
-    if(employee.getId() != 0){
-      employee.setUpdateDate(new Date());
-      iEmployeeService.updateObject(employee);
+    if(validateSave(employee)){
+      if(employee.getId() != 0){
+        employee.setUpdateDate(new Date());
+        iEmployeeService.updateObject(employee);
+      }else{
+        employee.setCreateDate(new Date());
+        employee.setActive(true);
+        iEmployeeService.insertObject(employee);
+      }
+      return "redirect:/employee_manager";
     }else{
-      employee.setCreateDate(new Date());
-      employee.setActive(true);
-      iEmployeeService.insertObject(employee);
+      if(employee.getId() != 0){
+        return "redirect:/update_employee?employeeId="+employee.getId();
+      }else{
+        return "redirect:/add_employee";
+      }
     }
-    return "redirect:/employee_manager";
   }
 
   @RequestMapping("/update_employee")
@@ -64,12 +73,15 @@ public class EmployeeController {
     model.addAttribute("employee",iEmployeeService.getObjectById(employeeId));
     model.addAttribute("departments",iDepartmentService.getAll());
     model.addAttribute("user",session.getAttribute("user"));
+    sentError(model);
     return "/html/Manager/employee/manager-employee-update";
   }
 
   @RequestMapping("/delete_employee")
   public String deleteEmployee(@RequestParam("employeeId")int employeeId) {
-    iEmployeeService.deleteObject(employeeId);
+    if(validateDelete(employeeId)){
+      iEmployeeService.deleteObject(employeeId);
+    }
     return "redirect:/employee_manager";
   }
 
@@ -82,21 +94,54 @@ public class EmployeeController {
     return "/html/Manager/department/manager-employeeOfDepartment";
   }
 
-//  private boolean validate(EmployeeEntity employee){
-//    boolean invalid = false;
-//    if(!checkRegularExpression("^[_A-Za-z0-9]+(\\.[_A-Za-z0-9]+)*@[A-Za-z0-9]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$",employee.getEmail())){
-//      errors.add("Email không đúng định dạng");
-//    }
-//    if(iEmployeeService.getAll().stream().map(EmployeeEntity::getEmail).collect(Collectors.toList()).contains(employee.getEmail())){
-//      errors.add("Email đã tồn tại");
-//    }
-//    if(!checkRegularExpression("^[0-9]+$", employee.getPhone())){
-//      errors.add("Không đúng định dạng số điện thoại");
-//    }
-//    if(iEmployeeService.getAll().stream().map(EmployeeEntity::getPhone).collect(Collectors.toList()).contains(employee.getPhone())){
-//      errors.add("Số điện thoại đã tồn tại");
-//    }
-//  }
+  private boolean validateSave(EmployeeEntity employee){
+    boolean valid = true;
+    if(!checkRegularExpression("^[_A-Za-z0-9]+(\\.[_A-Za-z0-9]+)*@[A-Za-z0-9]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$",employee.getEmail())){
+      errors.add("Email không đúng định dạng");
+      valid = false;
+    }
+    if(!checkRegularExpression("^[0-9]+$", employee.getPhone())){
+      errors.add("Không đúng định dạng số điện thoại");
+      valid = false;
+    }
+    if(employee.getBirthday() != null && employee.getBirthday().after(new Date())){
+      errors.add("Ngày sinh không thể lớn hơn ngày hiện tại");
+      valid = false;
+    }
+    if(employee.getId() != 0){
+      if(!iEmployeeService.getObjectById(employee.getId()).getEmail().equals(employee.getEmail())){
+        if(iEmployeeService.getAll().stream().map(EmployeeEntity::getEmail).collect(Collectors.toList()).contains(employee.getEmail())){
+          errors.add("Email đã tồn tại");
+          valid = false;
+        }
+      }
+      if(!iEmployeeService.getObjectById(employee.getId()).getPhone().equals(employee.getPhone())){
+        if(iEmployeeService.getAll().stream().map(EmployeeEntity::getPhone).collect(Collectors.toList()).contains(employee.getPhone())){
+          errors.add("Số điện thoại đã tồn tại");
+          valid = false;
+        }
+      }
+    }else{
+      if(iEmployeeService.getAll().stream().map(EmployeeEntity::getEmail).collect(Collectors.toList()).contains(employee.getEmail())){
+        errors.add("Email đã tồn tại");
+        valid = false;
+      }
+      if(iEmployeeService.getAll().stream().map(EmployeeEntity::getPhone).collect(Collectors.toList()).contains(employee.getPhone())){
+        errors.add("Số điện thoại đã tồn tại");
+        valid = false;
+      }
+    }
+    return valid;
+  }
+
+  private boolean validateDelete(int employeeId){
+    if(iEmployeeService.getAll().stream().map(EmployeeEntity::getId).collect(Collectors.toList()).contains(employeeId)){
+      return true;
+    }else{
+      errors.add("Nhân viên không tồn tại");
+      return false;
+    }
+  }
 
   private boolean checkRegularExpression(String regex, String str) {
     Pattern pattern = Pattern.compile(regex);
@@ -105,5 +150,17 @@ public class EmployeeController {
       return true;
     }
     return false;
+  }
+
+  private void sentError(Model model){
+    if(errors.size() > 0){
+      StringBuilder string = new StringBuilder();
+      for(String error : errors){
+        string.append(error);
+        string.append('\n');
+      }
+      model.addAttribute("error",string);
+      errors.clear();
+    }
   }
 }
